+++ src/TestSuite/TestTrait.php
@@ -18,7 +18,6 @@ namespace Tools\TestSuite;
 
 use BadMethodCallException;
 use Exception;
-use PHPUnit\Framework\Error\Error;
 use Throwable;
 
 /**
@@ -141,8 +140,6 @@ trait TestTrait
                     $e->getMessage()
                 ));
             }
-        } catch (Error $e) {
-            throw $e;
         }
 
         if (!isset($e)) {