+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 * all `_or_fail()` methods have been removed.
 
 ## 1.4 branch
+### 1.4.2
+* by default, `create_file()`, `dir_tree()` `is_writable_resursive()` and
+    `unlink_recursive()` functions can throw an exception. Added the third
+    `$ignoreErrors` parameter, which allows to ignore any errors and return a
+    default value;
+* `rmdir_recursive()` returns a boolean;
+* fixed bug for `dir_tree()` function on the filesystem root.
+
 ### 1.4.1
 * added `Exceptionist` class. All `_or_fail()` methods are now deprecated;
 * added `EventList::toArray()` method.